added a John Forkosh graph example and an all character test
The John Forkosh example was originally written as an ASCII art
("*" and " ") graph. He posted it to comp.lang.c in response to
a post about the DotDotDot project. The all character test
draws 256 cells of 2 wide and 4 tall, each with a different bit
pattern, thus showing the whole character set.